UPDATE ::: Moving date calculation to the BE from the DB due to sqlite limitations
git-svn-id: svn://107.170.10.188:8080/Semoss/dev@45887 ef391c1c-ee15-45f7-b1d6-dde3d27a2169

Former-commit-id: b24161399d2971b4142909e599eafb0e314227f7

+++ src/prerna/auth/utils/reactors/admin/AdminUserLockAccountWarningReactor.java
@@ -80,7 +80,7 @@ public class AdminUserLockAccountWarningReactor extends AbstractReactor {
 			}
 			for(Object[] emailInfo : listToEmail) {
 				String email = (String) emailInfo[0];
-				int daysSinceLastLogin = ((Number) emailInfo[1]).intValue();
+				long daysSinceLastLogin = ((Number) emailInfo[1]).longValue();
 				
 				Map<String, String> emailReplacements = SocialPropertiesUtil.getInstance().getEmailStaticProps();
 				emailReplacements.put(DAYS_TO_LOCK_REPLACEMENT, daysToLock + "");
